refactor(database): refactor database structure
