--- internal/database/sql.go
@@ -1,14 +1,0 @@
-package database
-
-import (
-	"context"
-	"database/sql"
-)
-
-// ISQL is a interface for database/sql
-type ISQL interface {
-	ExecContext(context.Context, string, ...interface{}) (sql.Result, error)
-	PrepareContext(context.Context, string) (*sql.Stmt, error)
-	QueryContext(context.Context, string, ...interface{}) (*sql.Rows, error)
-	QueryRowContext(context.Context, string, ...interface{}) *sql.Row
-}
