+++ internal/server/server.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/gofiber/fiber/v2"
 	"github.com/rasatmaja/zephyr-one/internal/config"
+	"github.com/rasatmaja/zephyr-one/internal/database"
 	"github.com/rasatmaja/zephyr-one/internal/handler"
 	"github.com/rasatmaja/zephyr-one/internal/logger"
 	"github.com/rasatmaja/zephyr-one/internal/middleware"
@@ -20,6 +21,7 @@ type App struct {
 	env        *config.ENV
 	middleware *middleware.App
 	utils      *utils.Registry
+	db         database.IRepository
 }
 
 // New is a function to initialize sever and its component
@@ -36,6 +38,8 @@ func New() *App {
 
 	// setup utils
 	utils := utils.New()
+
+	db := database.New()
 
 	// setup server
 	svr := fiber.New(
@@ -56,6 +60,7 @@ func New() *App {
 		env:        env,
 		middleware: mdlwre,
 		utils:      utils,
+		db:         db,
 	}
 }
